+++ event-sourcing/event-source/src/main/java/uk/gov/justice/services/eventsourcing/source/core/EventSourceNameExtractor.java
@@ -10,7 +10,6 @@ import javax.enterprise.inject.spi.InjectionPoint;
 @ApplicationScoped
 public class EventSourceNameExtractor {
 
-    @SuppressWarnings("unchecked")
     public String getEventSourceNameFromQualifier(final InjectionPoint injectionPoint) {
 
         return injectionPoint.getQualifiers().stream()