+++ event-sourcing/event-subscription-registry/src/main/java/uk/gov/justice/subscription/registry/RegistryException.java
@@ -2,7 +2,11 @@ package uk.gov.justice.subscription.registry;
 
 public class RegistryException extends RuntimeException {
 
-    public RegistryException(final String message){
+    public RegistryException(final String message) {
         super(message);
+    }
+
+    public RegistryException(final String message, final Throwable cause) {
+        super(message, cause);
     }
 }