+++ event-sourcing/event-subscription/src/main/java/uk/gov/justice/subscription/EventSourcesParser.java
@@ -5,10 +5,13 @@ import uk.gov.justice.subscription.domain.eventsource.EventSources;
 import uk.gov.justice.subscription.yaml.parser.YamlFileValidator;
 import uk.gov.justice.subscription.yaml.parser.YamlParser;
 
-import java.nio.file.Path;
+import java.net.URL;
 import java.util.Collection;
 import java.util.stream.Stream;
 
+/**
+ * Parse YAML URLs into {@link EventSource}s
+ */
 public class EventSourcesParser {
 
     private final YamlParser yamlParser;
@@ -19,13 +22,19 @@ public class EventSourcesParser {
         this.yamlFileValidator = yamlFileValidator;
     }
 
-    public Stream<EventSource> getEventSourcesFrom(final Collection<Path> paths) {
-        return paths.stream()
+    /**
+     * Return a Stream of {@link EventSource} from a Collection of YAML URLs
+     *
+     * @param urls the YAML URLs to parse
+     * @return Stream of {@link EventSource}
+     */
+    public Stream<EventSource> getEventSourcesFrom(final Collection<URL> urls) {
+        return urls.stream()
                 .flatMap(path -> parseEventSourcesFromYaml(path).getEventSources().stream());
     }
 
-    private EventSources parseEventSourcesFromYaml(final Path path) {
-        yamlFileValidator.validateEventSource(path);
-        return yamlParser.parseYamlFrom(path, EventSources.class);
+    private EventSources parseEventSourcesFromYaml(final URL url) {
+        yamlFileValidator.validateEventSource(url);
+        return yamlParser.parseYamlFrom(url, EventSources.class);
     }
 }