+++ event-sourcing/event-subscription/src/test/java/uk/gov/justice/subscription/YamlFileFinderTest.java
@@ -3,17 +3,23 @@ package uk.gov.justice.subscription;
 import static org.hamcrest.CoreMatchers.endsWith;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.verify;
 
-import java.nio.file.Path;
+import java.net.URL;
 import java.util.List;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.slf4j.Logger;
 
 @RunWith(MockitoJUnitRunner.class)
 public class YamlFileFinderTest {
+
+    @Mock
+    private Logger logger;
 
     @InjectMocks
     private YamlFileFinder yamlFileFinder;
@@ -21,20 +27,27 @@ public class YamlFileFinderTest {
     @Test
     public void shouldFindAllSubscriptionDescriptorsOnTheClasspathWhichHaveTheCorrectName() throws Exception {
 
-        final List<Path> paths = yamlFileFinder.getSubscriptionDescriptorPaths();
+        final List<URL> urls = yamlFileFinder.getSubscriptionDescriptorPaths();
 
-        assertThat(paths.size(), is(1));
+        assertThat(urls.size(), is(1));
 
-        assertThat(paths.get(0).toString(), endsWith("/subscription-descriptor.yaml"));
+        assertThat(urls.get(0).toString(), endsWith("/yaml/subscription-descriptor.yaml"));
     }
 
     @Test
     public void shouldFindAllEventSourcesOnTheClasspathWhichHaveTheCorrectName() throws Exception {
 
-        final List<Path> paths = yamlFileFinder.getEventSourcesPaths();
+        final List<URL> urls = yamlFileFinder.getEventSourcesPaths();
 
-        assertThat(paths.size(), is(1));
+        assertThat(urls.size(), is(1));
 
-        assertThat(paths.get(0).toString(), endsWith("/event-sources.yaml"));
+        assertThat(urls.get(0).toString(), endsWith("/yaml/event-sources.yaml"));
+    }
+
+    @Test
+    public void shouldLogFoundResources() throws Exception {
+        yamlFileFinder.getSubscriptionDescriptorPaths();
+
+        verify(logger).info("Found 1 resources on the classpath for yaml/subscription-descriptor.yaml");
     }
 }