+++ generators/messaging-adapter-generator/src/main/java/uk/gov/justice/subscription/jms/parser/SubscriptionWrapperFileParserFactory.java
@@ -4,10 +4,10 @@ import uk.gov.justice.maven.generator.io.files.parser.FileParserFactory;
 import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
 import uk.gov.justice.subscription.EventSourcesParser;
 import uk.gov.justice.subscription.SubscriptionDescriptorsParser;
-import uk.gov.justice.subscription.yaml.parser.YamlFileToJsonObjectConverter;
 import uk.gov.justice.subscription.yaml.parser.YamlFileValidator;
 import uk.gov.justice.subscription.yaml.parser.YamlParser;
 import uk.gov.justice.subscription.yaml.parser.YamlSchemaLoader;
+import uk.gov.justice.subscription.yaml.parser.YamlToJsonObjectConverter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -18,10 +18,15 @@ public class SubscriptionWrapperFileParserFactory implements FileParserFactory<S
         final YamlParser yamlParser = new YamlParser();
         final YamlSchemaLoader yamlSchemaLoader = new YamlSchemaLoader();
         final ObjectMapper objectMapper = new ObjectMapperProducer().objectMapper();
-        final YamlFileValidator yamlFileValidator = new YamlFileValidator(new YamlFileToJsonObjectConverter(yamlParser, objectMapper), yamlSchemaLoader);
+        final YamlFileValidator yamlFileValidator = new YamlFileValidator(new YamlToJsonObjectConverter(yamlParser, objectMapper), yamlSchemaLoader);
 
-        final EventSourcesFileParser eventSourcesFileParser = new EventSourcesFileParser(new EventSourcesParser(yamlParser, yamlFileValidator));
-        final SubscriptionDescriptorFileParser subscriptionDescriptorFileParser = new SubscriptionDescriptorFileParser(new SubscriptionDescriptorsParser(yamlParser, yamlFileValidator));
+        final EventSourcesParser eventSourcesParser = new EventSourcesParser(yamlParser, yamlFileValidator);
+        final PathToUrlResolver pathToUrlResolver = new PathToUrlResolver();
+
+        final EventSourcesFileParser eventSourcesFileParser = new EventSourcesFileParser(eventSourcesParser, pathToUrlResolver);
+
+        final SubscriptionDescriptorsParser subscriptionDescriptorsParser = new SubscriptionDescriptorsParser(yamlParser, yamlFileValidator);
+        final SubscriptionDescriptorFileParser subscriptionDescriptorFileParser = new SubscriptionDescriptorFileParser(subscriptionDescriptorsParser, pathToUrlResolver);
 
         return new SubscriptionWrapperFileParser(eventSourcesFileParser, subscriptionDescriptorFileParser);
     }