+++ generators/messaging-adapter-generator/src/test/java/uk/gov/justice/subscription/jms/it/JmsEndpointGenerationCustomIT.java
@@ -13,44 +13,66 @@ import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventFilter
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventListenerInterceptorChainProvider;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventEventValidationInterceptor;
 import uk.gov.justice.api.subscription.CustomEventListenerPeopleEventJmsListener;
-import uk.gov.justice.api.subscription.mapper.CustomEventListenerMediaTypeToSchemaIdMapper;
 import uk.gov.justice.raml.jms.it.AbstractJmsAdapterGenerationIT;
-import uk.gov.justice.schema.catalog.CatalogProducer;
-import uk.gov.justice.schema.service.SchemaCatalogService;
 import uk.gov.justice.services.adapter.messaging.DefaultJmsParameterChecker;
 import uk.gov.justice.services.adapter.messaging.DefaultJmsProcessor;
-import uk.gov.justice.services.common.configuration.ServiceContextNameProvider;
+import uk.gov.justice.services.adapter.messaging.DefaultSubscriptionJmsProcessor;
+import uk.gov.justice.services.adapter.messaging.JmsLoggerMetadataInterceptor;
+import uk.gov.justice.services.adapter.messaging.JsonSchemaValidationInterceptor;
+import uk.gov.justice.services.common.configuration.GlobalValueProducer;
+import uk.gov.justice.services.common.converter.ObjectToJsonValueConverter;
 import uk.gov.justice.services.common.converter.StringToJsonObjectConverter;
 import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
+import uk.gov.justice.services.common.util.UtcClock;
+import uk.gov.justice.services.core.accesscontrol.AccessControlFailureMessageGenerator;
+import uk.gov.justice.services.core.accesscontrol.AllowAllPolicyEvaluator;
+import uk.gov.justice.services.core.accesscontrol.DefaultAccessControlService;
+import uk.gov.justice.services.core.accesscontrol.PolicyEvaluator;
 import uk.gov.justice.services.core.cdi.LoggerProducer;
+import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
+import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
+import uk.gov.justice.services.core.dispatcher.EnvelopePayloadTypeConverter;
+import uk.gov.justice.services.core.dispatcher.JsonEnvelopeRepacker;
+import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
+import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
+import uk.gov.justice.services.core.envelope.EnvelopeInspector;
+import uk.gov.justice.services.core.envelope.EnvelopeValidationExceptionHandlerProducer;
+import uk.gov.justice.services.core.envelope.EnvelopeValidator;
+import uk.gov.justice.services.core.envelope.MediaTypeProvider;
+import uk.gov.justice.services.core.enveloper.Enveloper;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
-import uk.gov.justice.services.core.json.BackwardsCompatibleJsonSchemaValidator;
+import uk.gov.justice.services.core.extension.ServiceComponentScanner;
 import uk.gov.justice.services.core.json.DefaultFileSystemUrlResolverStrategy;
 import uk.gov.justice.services.core.json.DefaultJsonValidationLoggerHelper;
-import uk.gov.justice.services.core.json.FileBasedJsonSchemaValidator;
 import uk.gov.justice.services.core.json.JsonSchemaLoader;
-import uk.gov.justice.services.core.json.PayloadExtractor;
-import uk.gov.justice.services.core.json.SchemaCatalogAwareJsonSchemaValidator;
+import uk.gov.justice.services.core.json.JsonSchemaValidator;
 import uk.gov.justice.services.core.mapping.ActionNameToMediaTypesMappingObserver;
 import uk.gov.justice.services.core.mapping.DefaultMediaTypesMappingCache;
 import uk.gov.justice.services.core.mapping.DefaultNameToMediaTypeConverter;
-import uk.gov.justice.services.core.mapping.DefaultSchemaIdMappingCache;
+import uk.gov.justice.services.core.mapping.MediaType;
 import uk.gov.justice.services.core.mapping.MediaTypesMappingCacheInitialiser;
-import uk.gov.justice.services.core.mapping.MediaTypeToSchemaIdMapper;
 import uk.gov.justice.services.core.mapping.SchemaIdMappingCacheInitialiser;
 import uk.gov.justice.services.core.mapping.SchemaIdMappingObserver;
+import uk.gov.justice.services.core.requester.RequesterProducer;
+import uk.gov.justice.services.core.sender.SenderProducer;
 import uk.gov.justice.services.event.buffer.api.AllowAllEventFilter;
-import uk.gov.justice.services.generators.test.utils.interceptor.RecordingInterceptorChainProcessor;
+import uk.gov.justice.services.generators.test.utils.interceptor.EnvelopeRecorder;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.jms.DefaultEnvelopeConverter;
+import uk.gov.justice.services.messaging.jms.DefaultJmsEnvelopeSender;
 import uk.gov.justice.services.messaging.logging.DefaultJmsMessageLoggerHelper;
 import uk.gov.justice.services.messaging.logging.DefaultTraceLogger;
+import uk.gov.justice.services.subscription.SubscriptionManager;
+import uk.gov.justice.services.subscription.annotation.SubscriptionName;
 
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.annotation.Resource;
 import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 import javax.jms.Topic;
 
@@ -68,8 +90,6 @@ import org.junit.runner.RunWith;
 public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationIT {
     @Module
     @Classes(cdi = true, value = {
-            DefaultJmsProcessor.class,
-            RecordingInterceptorChainProcessor.class,
 
             CustomEventListenerPeopleEventEventFilter.class,
             CustomEventListenerPeopleEventEventFilterInterceptor.class,
@@ -83,42 +103,64 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
             CustomEventListenerExampleEventEventValidationInterceptor.class,
             CustomEventListenerExampleEventJmsListener.class,
 
-            ObjectMapperProducer.class,
+            RecordingJsonSchemaValidator.class,
+
+            ServiceComponentScanner.class,
+            RequesterProducer.class,
+            ServiceComponentObserver.class,
+            DefaultJmsProcessor.class,
+            SenderProducer.class,
+            DefaultJmsEnvelopeSender.class,
             DefaultEnvelopeConverter.class,
+            JsonSchemaValidationInterceptor.class,
+            JmsLoggerMetadataInterceptor.class,
+            DefaultJmsParameterChecker.class,
+            JmsAdapterToHandlerIT.TestServiceContextNameProvider.class,
+            JsonSchemaLoader.class,
             StringToJsonObjectConverter.class,
             DefaultJsonObjectEnvelopeConverter.class,
-            FileBasedJsonSchemaValidator.class,
-            JsonSchemaLoader.class,
+            ObjectToJsonValueConverter.class,
+            ObjectMapperProducer.class,
+            Enveloper.class,
+            AccessControlFailureMessageGenerator.class,
+            AllowAllPolicyEvaluator.class,
+            DefaultAccessControlService.class,
+            DispatcherCache.class,
+            DispatcherFactory.class,
+            EnvelopePayloadTypeConverter.class,
+            JsonEnvelopeRepacker.class,
+            PolicyEvaluator.class,
             LoggerProducer.class,
             AllowAllEventFilter.class,
-            DefaultJmsParameterChecker.class,
-            TestServiceContextNameProvider.class,
+            EmptySystemUserProvider.class,
+            SystemUserUtil.class,
+            BeanInstantiater.class,
+            UtcClock.class,
+            GlobalValueProducer.class,
+            EnvelopeValidationExceptionHandlerProducer.class,
             DefaultJmsMessageLoggerHelper.class,
             DefaultTraceLogger.class,
-            DefaultJsonValidationLoggerHelper.class,
+
             DefaultFileSystemUrlResolverStrategy.class,
+            DefaultJsonValidationLoggerHelper.class,
 
-            CustomEventListenerMediaTypeToSchemaIdMapper.class,
-            SchemaCatalogAwareJsonSchemaValidator.class,
-            PayloadExtractor.class,
             DefaultNameToMediaTypeConverter.class,
-
-            DefaultSchemaIdMappingCache.class,
-            SchemaIdMappingObserver.class,
-            MediaTypeToSchemaIdMapper.class,
-            BeanInstantiater.class,
-            MediaTypeToSchemaIdMapper.class,
-
-            CatalogProducer.class,
-            SchemaCatalogService.class,
-
             DefaultMediaTypesMappingCache.class,
             ActionNameToMediaTypesMappingObserver.class,
-
-            BackwardsCompatibleJsonSchemaValidator.class,
+            SchemaIdMappingObserver.class,
 
             MediaTypesMappingCacheInitialiser.class,
-            SchemaIdMappingCacheInitialiser.class
+            SchemaIdMappingCacheInitialiser.class,
+
+            SenderProducer.class,
+            MediaTypeProvider.class,
+            EnvelopeValidator.class,
+            EnvelopeInspector.class,
+            RequesterProducer.class,
+
+            DefaultSubscriptionJmsProcessor.class,
+            TestSubscriptionManagerProducer.class,
+            RecordingSubscriptionManager.class
     })
     public WebApp war() {
         return new WebApp()
@@ -126,7 +168,7 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
     }
 
     @Inject
-    RecordingInterceptorChainProcessor interceptorChainProcessor;
+    RecordingSubscriptionManager recordingSubscriptionManager;
 
     @Resource(name = "people.event")
     private Topic peopleEventsDestination;
@@ -142,7 +184,7 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
 
         sendEnvelope(metadataId, eventName, peopleEventsDestination);
 
-        final JsonEnvelope receivedEnvelope = interceptorChainProcessor.awaitForEnvelopeWithMetadataOf("id", metadataId);
+        final JsonEnvelope receivedEnvelope = recordingSubscriptionManager.awaitForEnvelopeWithMetadataOf("id", metadataId);
         assertThat(receivedEnvelope.metadata().id(), is(UUID.fromString(metadataId)));
         assertThat(receivedEnvelope.metadata().name(), is(eventName));
     }
@@ -155,17 +197,50 @@ public class JmsEndpointGenerationCustomIT extends AbstractJmsAdapterGenerationI
 
         sendEnvelope(metadataId, eventName, exampleEventDestination);
 
-        final JsonEnvelope receivedEnvelope = interceptorChainProcessor.awaitForEnvelopeWithMetadataOf("id", metadataId);
+        final JsonEnvelope receivedEnvelope = recordingSubscriptionManager.awaitForEnvelopeWithMetadataOf("id", metadataId);
         assertThat(receivedEnvelope.metadata().id(), is(UUID.fromString(metadataId)));
         assertThat(receivedEnvelope.metadata().name(), is(eventName));
     }
 
     @ApplicationScoped
-    public static class TestServiceContextNameProvider implements ServiceContextNameProvider {
+    public static class TestSubscriptionManagerProducer {
+
+        @Inject
+        RecordingSubscriptionManager recordingSubscriptionManager;
+
+        @Produces
+        @SubscriptionName
+        public SubscriptionManager subscriptionManager() {
+            return recordingSubscriptionManager;
+        }
+    }
+
+    @ApplicationScoped
+    public static class RecordingSubscriptionManager extends EnvelopeRecorder implements SubscriptionManager {
 
         @Override
-        public String getServiceContextName() {
-            return "test-component";
+        public void process(final JsonEnvelope jsonEnvelope) {
+            record(jsonEnvelope);
+        }
+    }
+
+    @ApplicationScoped
+    public static class RecordingJsonSchemaValidator implements JsonSchemaValidator {
+
+        private String validatedEventName;
+
+        @Override
+        public void validate(final String payload, final String actionName) {
+            this.validatedEventName = actionName;
+        }
+
+        @Override
+        public void validate(final String payload, final String actionName, final Optional<MediaType> mediaType) {
+            this.validatedEventName = actionName;
+        }
+
+        public String validatedEventName() {
+            return validatedEventName;
         }
     }
 }