+++ generators/messaging-adapter-generator/src/test/java/uk/gov/justice/subscription/jms/parser/EventSourcesFileParserTest.java
@@ -3,6 +3,7 @@ package uk.gov.justice.subscription.jms.parser;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.fail;
@@ -12,7 +13,9 @@ import static org.mockito.Mockito.when;
 import uk.gov.justice.subscription.EventSourcesParser;
 import uk.gov.justice.subscription.domain.eventsource.EventSource;
 
+import java.net.URL;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -26,29 +29,33 @@ import org.mockito.runners.MockitoJUnitRunner;
 public class EventSourcesFileParserTest {
 
     @Mock
-    EventSourcesParser eventSourcesParser;
+    private EventSourcesParser eventSourcesParser;
+
+    @Mock
+    private PathToUrlResolver pathToUrlResolver;
 
     @InjectMocks
     private EventSourcesFileParser eventSourcesFileParser;
 
     @Test
-    public void shouldReturnEventSources() {
-        final Path baseDir = mock(Path.class);
-        final Path resolvedPath = mock(Path.class);
+    public void shouldReturnEventSources() throws Exception {
+        final EventSource eventSource = mock(EventSource.class);
+        final Path baseDir = Paths.get("/yaml");
 
-        final Path eventSourcePath = mock(Path.class);
-        final Path subscriptionDescriptorPath = mock(Path.class);
+        final Path eventSourcePath = Paths.get("event-sources.yaml");
+        final Path subscriptionDescriptorPath = Paths.get("subscription-descriptor.yaml");
         final List<Path> pathList = asList(eventSourcePath, subscriptionDescriptorPath);
 
-        when(subscriptionDescriptorPath.endsWith("event-sources.yaml")).thenReturn(false);
-        when(eventSourcePath.endsWith("event-sources.yaml")).thenReturn(true);
+        final URL eventSourceUrl = baseDir.resolve(eventSourcePath).toUri().toURL();
+        final List<URL> urlList = singletonList(eventSourceUrl);
 
-        when(baseDir.resolve(eventSourcePath)).thenReturn(resolvedPath);
-        when(eventSourcesParser.getEventSourcesFrom(singletonList(resolvedPath))).thenReturn(Stream.of(mock(EventSource.class)));
+        when(pathToUrlResolver.resolveToUrl(baseDir, eventSourcePath)).thenReturn(eventSourceUrl);
+        when(eventSourcesParser.getEventSourcesFrom(urlList)).thenReturn(Stream.of(eventSource));
 
         final List<EventSource> expectedEventSourceDefinitions = eventSourcesFileParser.getEventSources(baseDir, pathList);
 
         assertThat(expectedEventSourceDefinitions.size(), is(1));
+        assertThat(expectedEventSourceDefinitions.get(0), is(eventSource));
     }
 
     @Test
@@ -58,7 +65,8 @@ public class EventSourcesFileParserTest {
         try {
             eventSourcesFileParser.getEventSources(baseDir, emptyList());
             fail();
-        } catch (final SubscriptionFileParserException e) {
+        } catch (final Exception e) {
+            assertThat(e, is(instanceOf(FileParserException.class)));
             assertThat(e.getMessage(), is("No event-sources.yaml files found!"));
         }
     }