+++ CHANGELOG.md
@@ -12,6 +12,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 - Updated test-utils version to 1.16.2
 - Updated file.service version to 1.16.0
 - Updated common bom version to 1.24.0
+- Change message adapter generator to use SubscriptionManager instead of the interceptor chain
 
 ## [4.0.0] - 2018-03-09